refactor: move multiplexed session handling to separate class
