+++ google-cloud-spanner/src/main/java/com/google/cloud/spanner/SpannerOptions.java
@@ -149,7 +149,6 @@ public class SpannerOptions extends ServiceOptions<Spanner, SpannerOptions> {
   private final DirectedReadOptions directedReadOptions;
   private final boolean useVirtualThreads;
   private final OpenTelemetry openTelemetry;
-  private final boolean useRandomChannel;
 
   enum TracingFramework {
     OPEN_CENSUS,
@@ -653,7 +652,6 @@ public class SpannerOptions extends ServiceOptions<Spanner, SpannerOptions> {
     attemptDirectPath = builder.attemptDirectPath;
     directedReadOptions = builder.directedReadOptions;
     useVirtualThreads = builder.useVirtualThreads;
-    useRandomChannel = builder.useRandomChannel;
     openTelemetry = builder.openTelemetry;
   }
 
@@ -763,7 +761,6 @@ public class SpannerOptions extends ServiceOptions<Spanner, SpannerOptions> {
     private boolean attemptDirectPath = true;
     private DirectedReadOptions directedReadOptions;
     private boolean useVirtualThreads = false;
-    private boolean useRandomChannel = false;
     private OpenTelemetry openTelemetry;
 
     private static String createCustomClientLibToken(String token) {
@@ -828,7 +825,6 @@ public class SpannerOptions extends ServiceOptions<Spanner, SpannerOptions> {
       this.attemptDirectPath = options.attemptDirectPath;
       this.directedReadOptions = options.directedReadOptions;
       this.useVirtualThreads = options.useVirtualThreads;
-      this.useRandomChannel = options.useRandomChannel;
     }
 
     @Override
@@ -1325,11 +1321,6 @@ public class SpannerOptions extends ServiceOptions<Spanner, SpannerOptions> {
       return this;
     }
 
-    public Builder setUseRandomChannel(boolean useRandomChannel) {
-      this.useRandomChannel = useRandomChannel;
-      return this;
-    }
-
     @SuppressWarnings("rawtypes")
     @Override
     public SpannerOptions build() {
@@ -1570,10 +1561,6 @@ public class SpannerOptions extends ServiceOptions<Spanner, SpannerOptions> {
   @BetaApi
   public boolean isUseVirtualThreads() {
     return useVirtualThreads;
-  }
-
-  public boolean isUseRandomChannel() {
-    return useRandomChannel;
   }
 
   /** Returns the default query options to use for the specific database. */
